Add initial xcodebuild.vim implementation
This is the first working version of this plugin. It allows you to
dynamically find the project or workspace, as well as dynamically
finding the first available scheme (looking only in the project).

It also has support for xcpretty, and dynamically adds it to the end of
the command if it's available.