Dynamically decide which sdk to test against
If we're testing iOS apps, we want to specify the simulator, which means
we need to know if we're trying to test an iOS app or an OS X app.

In order to do so, we can just look at the output of
`xcodebuild -showBuildSettings` for the chosen project/scheme. If we see
a setting for `CORRESPONDING_SIMULATOR_SDK_NAME`, then we know it's an
iOS project, and can target the `iphonesimulator` sdk. To be safe, if we
_don't_ find that setting, we're pointing at the `macosx` sdk
explicitly.