Use `map` to create the cli_args
Turns out, vimscript has a _really_ wacky version of `map` for lists and
dictionaries. You can pass a list or a dict, and a string that will get
interpreted. Inside the string, you get access to `v:val` which is the
current value. The expression should return a string.

We then pass that through `join`, and prepend the whole thing with a
leading space.

Hat tip to @gabebw for pointing me in this direction. I'm all about them
functors.