Use custom build directory
Instead of using the build directory defined by Xcode, we can override
it and build in the current directory. The benefit here is that in the
_future_, we can use this command to build the project and then use
`simctl` to load the app in the iOS simulator. That's not something we
can do using the default build location (AFAIK).

In addition, we can let users customize this location with a global
variable. By default, we'll build in `./build`, but if users want to
send the build to a different location, they can do that easily.