Always pass the sdk flag
Turns out that if we don't pass the sdk info, it builds properly, but if
we wanted to do something crazy like load the built app into a
simulator, it'd fail because of incorrect architectures. So instead, we
should always build for the chosen SDK.

AFAIK, there's no downside to doing this.