Don't override build directory
We were overriding this in order to support a future feature (the
ability to run an app from Vim), but this also broke the ability for
`xcodebuild` to find implicit dependencies.

I think it's worthwhile to remove this from the codebase and try to find
another option for supporting a run command when we get to that.