Add more shell escaping when passing args
Previously we were just passing the project, workspace, and scheme to
the command line completely unescaped. This isn't super safe. Instead,
we should pass all of these through our `cli_args` function to add
quotes and a leading space for everything that's passed to the command
line.

Hat tip to https://github.com/gfontenot/vim-xcodebuild/pull/34 for the
original implementation.