Rename project to xcode.vim
This project does more than just wrap xcodebuild, and it's only going to
diverge more as time goes on. We should embrace that instead of feeling
limited by it, and rename the project to xcode.vim, which reflects the
fact that this is a plugin for working alongside Xcode from within Vim.