Allow users to specify a file to open with XOpen
Instead of opening the full project every time, it can be really useful
to open a specific file in Xcode. For example, if you're having trouble
debugging a specific class and you want Xcode's faster feedback loop, or
if you want to open a storyboard in Xcode for editing.

This modifies the `XOpen` command so that it can take an optional
argument. If that argument is provided, it will open the specified file
in Xcode. If the argument is omitted, it will open the entire project.

Note that Xcode itself is fairly smart about this. If you already have
the project open, for example, it will open the specified file _inside
the open project_, instead of opening a new window. This is handy.