Allow users to change the project
Previously, we were naively assuming a single xcode project at the root of the
project repo. Obviously, this was a dumb assumption. If a user had multiple
projects at the same level, we'd end up sending all of them to the commands
expecting a project name. This broke the commands and it wasn't possible to
recover.

To fix this, we can follow the same pattern we used for selecting the schemes.
We can check to see if the user has selected a project on first run, and if
not we can supply a reasonable default. As with the scheme, we'll default to
the first one we see.