Allow users to provide default projects/schemes
Previously, if xcode.vim was selecting the wrong scheme or project, the only
thing you could do was to correct the selection for the lifetime of the
current session. This meant that you'd have to reset those values every time
you opened vim. This is fairly annoying.

To fix this, we can first check for the existence of a global variable before
choosing our own default. If it's there, we can then use that instead of
trying to infer which to use. If it isn't we can still fall back to a sane
default.