Lowercase command names
Instead of using commands like XBuild, we should follow the conventions of
other plugins such as Fugitive and use commands like Xbuild

I'm also simplifying XSelectProject and XSelectScheme to Xproject and Xscheme,
respectively. This is partially because Xselectproject looks like shit, and
partially because vim favors brevity in cases like this.