Allow the specification of the workspace to use
Like with projects, we've been pretty naive about workspaces thus far.
Specifically, we've been assuming there is only one workspace, and we haven't
been giving the user hooks into specifying which workspace to use if that
isn't the case.

To fix this, we'll copy the treatment we're using for projects and use it for
workspaces. We'll default to the first one we find, but we'll expose a
`Xworkspace` command for setting it on the fly, as well as a
`xcode_workspace_file` variable to set externally.