Allow workspaces to be the project at the root
Previously, we were only looking for projects at the root. This is probably
fine most of the time, but it's easy to see that there _might_ be a time when
projects are nested in a subdirectory and there is instead a workspace at the
root. If that's the case, we should be fine.

To fix this issue, we should first check for projects and then if there aren't
any, we should check for a workspace file. If neither exist, _then_ we should
throw an error.