Pass workspace to helper scripts if available
Previously, we were passing the project file to the helper scripts and
completely disregarding the workspace. This doesn't actually make much sense,
and it would make more sense if we just pass the workspace if available.

To make this happen, we'll pass the properly formatted target string directly
to the helper script and pass the already formatted arguments to `xcodebuild`.

We'll also take this opportunity to make sure we're formatting our helper
scripts consistently and using `xcrun` before our `xcodebuild` commands.