Run iOS apps
This introduces the Xrun command, which allows users to run iOS apps from the
command line. It works by using xcodebuild to determine the build settings for
the chosen workspace/project/scheme and finding the app inside the build
directory for that setup. Once it sees the app, it can then find the bundle
identifier for the app and launches a simulator for the specified device type
(currently hardcoded to an iPhone 6s) and installs and launches the app.

This is a little naive, but it should work as an MVP, and removes one more
thing I previously still needed to move over to Xcode for.