Add the ability to pass actions to Xbuild
It's reasonable to think that users might want to run commands other than just
a simple `build` command. To make this easy, we can change the Xbuild command
so that it accepts actions to perform, defaulting to `build` if none are
provided. This means if you need to you could run `:Xbuild clean build<CR>`
and it'd clean and build. It also means you could run any action supported by
xcodebuild itself. I suppose it also means you could abuse the command to pass
_any_ options to xcodebuild that you want, but we'll autocomplete to the
actions to try guiding people in the right direction.