Xrun should also build the app
This is obvious in hindsight, but instead of forcing users to build manually
and then run as a separate step we should instead build _and_ run when the
user invokes `Xrun`. This _greatly_ simplifies the UI and is much more
intuitive.

Note that this meant refactoring the commands so that we're always passing the
preferred simulator from the top down. For _most_ of the commands, it doesn't
really matter, but for `Xrun` we want to make sure that we're building and
running with the same settings consistently.