Add neovim support
Neovim removed `!` as a command (I believe), and replaced it with a built-in
terminal emulator. This adds support for Neovim in the most basic way
possible, which is to default to `terminal` instead of `!` when we're using
Neovim.

We're also switching out the runner for `s:switch` because of the same issue.