Refactor neovim support
I meant to merge this as a part of b4ecbda but _totally_ forgot. Dumb.

Anyway, this refactors the neovim support so that it lives in one place. In
this case, we're defining our system runner based on if we're inside neovim or
not and then re-using that system runner for the default runner as well as for
the switch command.