Tell globpath to return a list
Turns out, there's an optional argument you can pass to `globpath` to get it
to return a native list, instead of returning a string that you then need to
parse. From `:h globpath`:

> When {list} is present and it is |TRUE| the result is a List
> with all matching files. The advantage of using a List is, you
> also get filenames containing newlines correctly. Otherwise
> the result is a String and when there are several matches,
> they are separated by <NL> characters.  Example:
>     :echo globpath(&rtp, "syntax/c.vim", 0, 1)

Unfortunately, we need to pass the also-optional `{nosuf}` argument as well.
The default for that is FALSE, so we'll just go ahead and pass that along as
well.