Improve project detection
Previously, we were only using the projects/workspaces in the root of
the project repo to determine if we could run our Xcode commands. This
made sense for the most part but also meant that if you explicitly
declared a path to an Xcode project/workspace and it _wasn't_ in the
root, we wouldn't let you run these commands. This made working with
non-standard project layouts (like those in React Native projects)
difficult to work with.

To fix this, we can refactor our project/workspace detection to take
the explicit file paths into account. If you're explicitly defining a
project path and it doesn't exist, I feel like that's your fault and you
deserve what you get.