Open specified project if available
Instead of opening `.` with Xcode (which just happens to open a project
in the current directory), we can be more explicit and open the chosen
project or workspace file. This has the nice benefit of explicitly
opening the project that users have specified in their config. This
means that this command will now work properly if the user is trying to
work with a project that lives in a nested subdirectory (like is the
case with React Native projects)