Only list the first build directory
At some point, Xcode started putting this setting in multiple places. This
caused the subsequent commands to fail because we were trying to use `find`
with a wacky string that included newlines.

To fix this, we can use the `-m` flag for grep to only display the first
result.

This is probably fragile, but progress is progress and this entire thing is a
horrifying hack anyway.