Only find simulator builds
Turns out it's pretty common for device builds to appear before simulator
builds with `find`. That means that this would install a device build into the
simulator and then try to launch it. Which failed miserably.

The simplest fix for this is to only show results for the simulator, and use
that path.