Add configuration for ignoring schemes
This adds a new `g:xcode_scheme_ignore_pattern` global variable that, if set,
will determine which schemes to hide when listing/finding schemes. This is
useful for cases where you have demo/example schemes that you don't want
included in the list of available schemes. If this isn't set, the old behavior
is maintained